Add sync: true for Redis, Kafka, and logger.close
I noticed "sync: true" is set in spec/spec_helper.rb for testing, so perhaps it's required when sending to those systems. I add here primarily for the benefit of quick testers, not as an optimization.

I also added an explicit logger.close at the end of the example script, which achieves the same effect by implicitly flushing output. I briefly considered using a method with a yield statement so writes could flush themselves when they go out of scope, but I'll leave that as a comment / suggestion for down the road.