Flush log messages when exiting
The current implementation of message buffering puts messages into a queue, and flushes them every five seconds. If the program exits, any messages in the buffer will be lost.

This fixes the issue by automatically flushing the buffer when the program exits. This is done by registering an `at_exit` block which does a final flush, which will wait for all messages to be flushed. If an exception occurs, it will be logged and ignored.

This option is configurable with the new option `buffer_flush_at_exit`, which now defaults to `true`.