Fix integration with ActiveJob
`ActiveJob` relies on access to `Logger#formatter` so that it can call `logger.formatter.current_tags`. In the current implementation, `MultiLogger#formatter` is always `nil`. This fixes the issue by ensuring that the base implementation is called when a formatter is set. This results in the instance variable `@formatter` being set, which fixes access to `Logger#formatter`.

Fixes https://github.com/dwbutler/logstash-logger/issues/112